changed weight type

+++ main.py
@@ -223,7 +223,7 @@ def main():
 
     model_ft = model_ft.to(device)
 
-    criterion = torch.nn.CrossEntropyLoss(weight=torch.tensor(weight)).to(device)
+    criterion = torch.nn.CrossEntropyLoss(weight=torch.tensor(weight,dtype=torch.float)).to(device)
 
     # Observe that all parameters are being optimized
     optimizer_ft = torch.optim.Adam(model_ft.parameters(), lr=args.lr)
